requirements.txt
Pillow==11.0.0
typer==0.12.5
typing_extensions==4.12.2
